my_fastapi_project/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic>=2.7.4
pydantic-settings>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
faiss-cpu==1.13.2
python-dotenv==1.0.0
httpx==0.25.1
pymongo==4.6.0
tqdm==4.66.1
